requirements.txt
duckduckgo_search==5.3.0b4
openai-python
streamlit